config: also save next-server in the state file
The early boot tooling gets the root-path from our state file due to a
better way to do that. However, when booting with NFS root, the root path
alone is not sufficient; the server address is communicated via the
next-server option. Save that one in the state file as well.