device: don't start connectivity checking on unconfigured devices
If the interface has no carrier, no addresses or no routes there is no
point in starting a connectivity check on it because we already know
that it will fail. Moreover, doing the check on a device without
routes causes the addition of a negative entry in the ARP table for
each of the addresses associated with the connectivity check
host; this can lead to poor network performances.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/issues/181