cli: accept " " as "gsm.apn" to indicate setting to ""
Until just recently, libnm would reject "" as an invalid setting
for "gsm.apn".

%NULL was allowed, and indicated to set the APN in ModemManager to
empty (""). The behavior for $NULL will change to allow lookup the
APN in the mobile-broadband-provider-info. Anyway, "" was kind of
a valid value all along, in fact with the same meaning as %NULL.
(note that oFono implementation migth be different here as it treats
%NULL and "" not the same).

Anyway, since libnm now allows to explicitly set the value to "",
also let nmcli support that. The problem is, that nmcli coerces
"" to %NULL, to reset the value. So, in nmcli, "" already has a
different meaning.

Extend this, so that " " means "", and "" continues to mean %NULL.

Note that the property setter of "gsm.apn" already strips whitespaces
from the APN. So, while we hijack " " to mean "", it was never (and
still is not) possible to set "gsm.apn" to " ". So we can still express
all values that we could previously.

Of course, also the getter needs to be adjsted.