fixup! initrd: add devicetree support
There is a g_return_*() in betwee, which would leak the connection.
Ok, that's an assertion (not supposed to happen anyway), so a leak
in that case would be OK (though easily avoidable).

But nmi_dt_reader_parse() is 250 LOC of parsing. For parsing, it is most
beneficial when every bit of resource/memory is owned by a variable on
the stack, and you can abort (return) at any point. This simplifies
review, because you just check that every resource that gets allocated
is owned by somebody.