fixup! initrd: add devicetree support
I dislike that dt_get_property() does two different things,
depending on whether @contents is requested.

If contents was not requested, it would return success/failure
based on whether the file existed. If contents was requested,
it would return success/failure based on whether it could read
the file. That's confusing, the function should always do the same
thing, and if somebody doesn't care for the contents, just not return
it.