fixup! initrd: add devicetree support
Use nm_utils_strsplit_set_full(). The resulting list of tokens
is an array of "const char *", which lets the compiler warn if
we would modify any of the tokens. That is nice, because we
assign the tokens to other variables (s_ipaddr), so knowing that
tokens are immutable (with compiler's help) makes that easier to
follow.

(also, on a minor note it gets the job done with fewer allocations).