client: Use GTask to initialize and populate the objects in a thread
NMClient takes a lot of time in the initialization phase to go through the
objects and initialize them, even if this is using an async initialization,
part of the objects initializations happens actually in the finish function.

To ensure that async init is really not-blocking, use a GTask thread to perform
the objects initialization.

Since now the objects's async init might not run in the main thread, and since
during this phase there might be signal emissions, we force this to happen in
the main context.

Fixes https://gitlab.freedesktop.org/NetworkManager/NetworkManager/issues/239