libnm-core: return NULL from boxed properties getters when unset
Boxed properties have a default value of NULL. If the getter returns
anything different from NULL when the value is not set, for example an
empty GHashTable or strv, then a nm_setting_diff() between a newly
created setting and NULL with
NM_SETTING_COMPARE_FLAG_DIFF_RESULT_NO_DEFAULT will report the
property as changed even if nothing really changed.

Update the boxed properties to return NULL when they are not set and
add a test for that.