core: add better logging for autoconnect with a new log domain
Sometimes users report problems about profiles not autoconnecting and
with the current logging it's difficult to understand where the
problem is. Since all profiles must be evaluated when a device becomes
available for autoconnection, adding by default a log message for each
of them would be probably too verbose, even for TRACE level.

Add a new 'autoconnect' log domain that is disabled by default and is
useful to debug autoconnect problems, and use it to log why profiles
can't autoconnect.