license: relicense "libnm-core/nm-vpn-dbus-interface.h" under LGPL-2.1+
We want to consistently use LGPL. Also, this file is part of libnm, which
requires its sources to be LGPL.

Also, this header file is supposed to be usable without libnm (e.g. for QT).

  $ git shortlog -n -s e055bdbbc3e7bb3ff3bc73860c4479f15bbfe7b2 -- libnm-core/nm-vpn-dbus-interface.h libnm-core/NetworkManagerVPN.h include/NetworkManagerVPN.h
      14    Dan Williams
       9    Dan Winship
       5    Thomas Haller
       4    Lubomir Rintel
       2    Jiří Klimeš
       2    Tambet Ingo
       1    Atul Anand
       1    Beniamino Galvani
       1    Evan Broder
       1    Zephaniah E. Loss-Cutler-Hull

This is a large number of contributors. Changing the license here is
problematic. However, there is also the stance that simple header files
are not copyrightable. Maybe more importantly, the vast majority of
contributions come from people who would agree with this change and/or
were employed by Red Hat at the time (possibly the copyright lies with
Red Hat).

See contributions done by not the usual suspects:

  $ git log --stat -p --no-walk $(git log --pretty='%H:%an <%ae>' e055bdbbc3e7bb3ff3bc73860c4479f15bbfe7b2 -- libnm-core/nm-vpn-dbus-interface.h libnm-core/NetworkManagerVPN.h include/NetworkManagerVPN.h | grep '\(dcbw\|bgalvani\|jklimes\|thaller\)@redhat.com\|danw@gnome.org\|lkundrak@v3.sk' -v | sed 's/:.*//') -- libnm-core/nm-vpn-dbus-interface.h libnm-core/NetworkManagerVPN.h include/NetworkManagerVPN.h