libnm: avoid leaking GMainContext for sync initialization after context-busy-watcher quits
When passing a destroy notify to g_dbus_connection_signal_subscribe(),
that callback gets invoked as an idle handler of the associated
GMainContext. That caused to have yet another source attached to the
context after the NMClient gets destroyed. Especially with synchronous
initialization that is bad, because we might destroy the
context-busy-watcher first. That results in removing the integration
of the inner GMainContext in the caller's context, and thus we leak the
entire context indefinitely.

Avoid that leak by not (indirectly) passing a cleanup function to
g_dbus_connection_signal_subscribe().

Fixes: ce0e898fb476 ('libnm: refactor caching of D-Bus objects in NMClient')