libnm: add nm-prop-meta for static meta data about properties and use it for bridge
We either keep duplicating the meta data all over the place (e.g.
for bridge options), or we don't have any useful meta data about
properties (so we cannot handle them generically).

Add "nm-prop-meta.h" that exposes meta data about settings
and properties.

Currently only used for Bridge master and port, but the naming
and layout is intended that it could be used for property meta
data for other settings.