libnm: add public library "libnm-gpl"
Historically, keyfile handling was part of core. It is thus GPL-2.0+
licensed. Quite some time ago, the parsing code was moved from core to
a distinct internal library (of course, keeping the license).

We would like to expose keyfile functionality via libnm, but due to the
licensing conflict, there are problems. We would either have to rewrite
the library from scratch or follow through with the relicensing. For
now, expose the functionality as separate GPL-2.0+ library, so that it
can be used by tools.

We don't add a special libnm-keyfile.so library, instead a general
libnm-gpl.so file that provides code that cannot be part of libnm.so due
to licensing. Hopefully one day we resolve the issue for keyfile and the
keyfile code could become part of libnm.so.

This will finally expose the documented format for loading and storing
connections to file.

At this point the API is still empty. That will be done next.