libnm: drop unused internal API _nm_setting_option_reset_from_hash() and _nm_setting_option_to_gvalue()
This was intended if the gendata hash should be converted
to/from a GValue/GHashTable. This would have been used, if
we also would have added a GObject property that exposes
the hash. But that was never done (at least not for NMSettingEthtool
and not yet).

This code is not used. If you ever need it, revert the patch
or implement it anew.